t migrate combination tests to junit 5
