+++ approvaltests-tests/src/test/java/org/approvaltests/combinations/CombinationTest.test1Parameter.approved.txt
