+++ approvaltests-tests/src/test/java/org/approvaltests/combinations/CombinationTest.testCombinations.approved.txt
