+++ approvaltests-tests/src/test/java/org/approvaltests/combinations/CombinationTest.testLockDown.approved.txt
