+++ approvaltests-tests/src/test/java/org/approvaltests/combinations/CombinationTest.testPassMethod.approved.txt
